COR-397: refactor redux slice
