+++ frontend/apps/pwa/src/features/former/FormerContainer.tsx
@@ -6,21 +6,20 @@ import { useAppDispatch, useAppSelector } from '../../app/hooks';
 import { fetchDatabases } from '../../reducers/database';
 import { fetchFolders } from '../../reducers/folder';
 import { fetchForms } from '../../reducers/form';
-import {
-  formerActions,
-  formerGlobalSelectors,
-  postForm,
-} from '../../reducers/former';
+import former from '../../reducers/Former';
+import { formerGlobalSelectors } from '../../reducers/Former/former.selectors';
+import { postForm } from '../../reducers/Former/former.reducers';
 
 import { Former } from './Former';
 
 export const FormerContainer: FC = () => {
   const dispatch = useAppDispatch();
   const history = useHistory();
+  const { actions } = former;
 
   // load data
   useEffect(() => {
-    dispatch(formerActions.reset());
+    dispatch(actions.reset());
     dispatch(fetchDatabases());
     dispatch(fetchFolders());
     dispatch(fetchForms());
@@ -46,18 +45,18 @@ export const FormerContainer: FC = () => {
     const search = new URLSearchParams(location.search);
     const databaseId = search.get('databaseId');
     if (databaseId) {
-      dispatch(formerActions.setDatabase({ databaseId }));
+      dispatch(actions.setDatabase({ databaseId }));
     }
     const folderId = search.get('folderId');
     if (folderId) {
-      dispatch(formerActions.setFolder({ folderId }));
+      dispatch(actions.setFolder({ folderId }));
     }
   }, [dispatch, location]);
 
   const setFormName = useCallback(
     (formName: string) => {
       if (form) {
-        dispatch(formerActions.setFormName({ formId: form.formId, formName }));
+        dispatch(actions.setFormName({ formId: form.formId, formName }));
       }
     },
     [dispatch, form],
@@ -65,8 +64,8 @@ export const FormerContainer: FC = () => {
 
   const setSelectedField = useCallback(
     (fieldId: string | undefined) => {
-      dispatch(formerActions.selectField({ fieldId }));
-      dispatch(formerActions.resetFormErrors());
+      dispatch(actions.selectField({ fieldId }));
+      dispatch(actions.resetFormErrors());
     },
     [dispatch],
   );
@@ -74,7 +73,7 @@ export const FormerContainer: FC = () => {
   const addField = useCallback(
     (kind: FieldKind) => {
       if (form) {
-        dispatch(formerActions.addField({ formId: form.formId, kind }));
+        dispatch(actions.addField({ formId: form.formId, kind }));
       }
     },
     [dispatch, form],
@@ -82,101 +81,99 @@ export const FormerContainer: FC = () => {
 
   const setFieldOption = useCallback(
     (fieldId: string, i: number, value: string) => {
-      dispatch(formerActions.setFieldOption({ fieldId, i, value }));
+      dispatch(actions.setFieldOption({ fieldId, i, value }));
     },
     [dispatch],
   );
 
   const addOption = useCallback(
     (fieldId: string) => {
-      dispatch(formerActions.addOption({ fieldId }));
+      dispatch(actions.addOption({ fieldId }));
     },
     [dispatch],
   );
 
   const removeOption = useCallback(
     (fieldId: string, i: number) => {
-      dispatch(formerActions.removeOption({ fieldId, i }));
+      dispatch(actions.removeOption({ fieldId, i }));
     },
     [dispatch],
   );
 
   const cancelField = useCallback(
     (fieldId: string) => {
-      dispatch(formerActions.cancelFieldChanges({ fieldId }));
+      dispatch(actions.cancelFieldChanges({ fieldId }));
     },
     [dispatch],
   );
 
   const setFieldRequired = useCallback(
     (fieldId: string, required: boolean) => {
-      dispatch(formerActions.setFieldRequired({ fieldId, required }));
+      dispatch(actions.setFieldRequired({ fieldId, required }));
     },
     [dispatch],
   );
 
   const setFieldIsKey = useCallback(
     (fieldId: string, isKey: boolean) => {
-      dispatch(formerActions.setFieldIsKey({ fieldId, isKey }));
+      dispatch(actions.setFieldIsKey({ fieldId, isKey }));
     },
     [dispatch],
   );
 
   const setFieldName = useCallback(
     (fieldId: string, name: string) => {
-      dispatch(formerActions.setFieldName({ fieldId, name }));
+      dispatch(actions.setFieldName({ fieldId, name }));
     },
     [dispatch],
   );
 
   const setFieldDescription = useCallback(
     (fieldId: string, description: string) => {
-      dispatch(formerActions.setFieldDescription({ fieldId, description }));
+      dispatch(actions.setFieldDescription({ fieldId, description }));
     },
     [dispatch],
   );
 
   const setFieldReferencedDatabaseId = useCallback(
     (fieldId: string, databaseId: string) => {
-      dispatch(
-        formerActions.setFieldReferencedDatabaseId({ fieldId, databaseId }),
-      );
+      dispatch(actions.setFieldReferencedDatabaseId({ fieldId, databaseId }));
     },
     [dispatch],
   );
 
   const setFieldReferencedFormId = useCallback(
     (fieldId: string, formId: string) => {
-      dispatch(formerActions.setFieldReferencedFormId({ fieldId, formId }));
+      dispatch(actions.setFieldReferencedFormId({ fieldId, formId }));
     },
     [dispatch],
   );
 
   const openSubForm = useCallback(
     (fieldId: string) => {
-      dispatch(formerActions.openSubForm({ fieldId }));
+      dispatch(actions.openSubForm({ fieldId }));
     },
     [dispatch],
   );
 
   const saveField = useCallback(
     (fieldId: string) => {
-      dispatch(formerActions.resetFormErrors());
-      dispatch(formerActions.selectField({ fieldId: undefined }));
+      dispatch(actions.resetFormErrors());
+      dispatch(actions.selectField({ fieldId: undefined }));
     },
     [dispatch],
   );
 
   const saveForm = useCallback(async () => {
     if (ownerForm) {
-      dispatch(formerActions.saveForm());
+      dispatch(actions.saveForm());
     } else if (formDefinition) {
-      dispatch(formerActions.resetFormErrors());
+      dispatch(actions.resetFormErrors());
       try {
         const data = await dispatch(postForm(formDefinition)).unwrap();
         history.push(`/browse/forms/${data.id}`);
       } catch (e: any) {
-        dispatch(formerActions.setFormErrors({ errors: e?.details?.causes }));
+        dispatch(actions.setFormErrors({ errors: e?.details?.causes }));
       }
     }
   }, [dispatch, formDefinition, ownerForm]);
