+++ frontend/apps/pwa/src/types/errors.ts
@@ -1,0 +1,7 @@
+export type ApiErrorDetails = {
+  field: string;
+  message: string;
+  reason: string;
+};
+
+export type ErrorMessage = Record<string, any>;
